Enhance output filename and working directory
The script ran in cwd and ended by deleting *.jpg, which isn't good. Now
it creates a working directory and all temporary files go there,
avoiding the possibility of existing files being overwritten. The
temporary directory is removed after the output file is copied out.

And it is not possible to specify the output filename on the command
line.